Load current function's arguments, not registered function's
Instead of loading arguments based on the names of arguments of
registered functions, load arguments based on the names of the arguments
of the current function when we're generating IR that checks the first
argument's type and dispatches to the correct implementation.

This also adds a test to make sure that we're using the argument names
of the current function, not the registered implementation.