Remove test with different type annotation and register argument
The testTypeAnnotationsDisagreeWithRegisterArgument test fails on CI
runs because the compiled code raises a TypeError when trying to pass a
value with the incorrect type.

We could try and support this by skipping the type check in this case or
changing the type of the first argument to the type of the register
argument, but this will probably turn into a mypy error at some point
anyway, so it's not really worth the effort to keep this test around.