Use non-native calls to registered implementations
Registered implementations are always decorated because they always have
a register decorator, and we can't make native calls to decorated
functions because otherwise we wouldn't be able to use the function
emitted from the decorator (which would be a Python callable object).

Previously, we removed the register decorator on those implementations
and treated them as non-decorated functions, but that strategy stops
working as soon as you add another decorator, so this just treats those
implementations as regular decorated functions and uses a non-native
call for those functions.