Avoid compilation error when using ABC
When trying to use the Mapping ABC as the dispatch type for a registered
implementation, we previously tried to load the type for the isinstance
check as `typing.Mapping` even if it was imported from
`collections.abc`, causing a compilation error due to the fact that we
hadn't defined CPyModule_typing.

To fix that, this loads the type from the globals dict instead for most
types, and using the types in builtin_names for builtin types, which
won't be present in the globals dict.