Show error when applying decorator after registering
Correctly supporting decorators being applied to functions after they've
been registered complicates the implementation because we wouldn't be
able to generate a call to the function that's been registered when
dispatching. Instead, we would have to:
- run all the decorators below the register decorator (we already do
  this)
- register the function at runtime by storing what the function looks
  like after those decorators have been run
- run the rest of the decorators

Doing that makes the implementation more complicated and having
decorators run after registering probably isn't common anyways, so we
just show an error.