Avoid error for creating underscore object
Don't show an error about calling a function named '_' directly when '_'
refers to a class named '_' and we're actually creating an object.
Having a class named '_' is probably rare and using it directly is
probably even rarer, but the other parts of the PR only deal with
functions, not classes, so we probably should keep that consistent.