Handle adjacent decorated underscore functions in semanal
Instead of turning adjacent decorated underscore functions into a series
of FuncDefs instead of one OverloadedFuncDef in the parser, handle
OverloadedFuncDefs named '_' in semanal. That prevents us from needing
to make that change in fastparse2, and move the handling of the
underscore function case closer to where the error would be emitted.