Fix handling of default arguments in generated glue methods
Currently these mostly segfault.

The main tricky case here is with *args and **kwargs (which we need to
deal with always when generating pycall glue and also when a native
function takes them), though we also need to take care to coerce error
values when forwarding optional arguments.

I also accidentally fixed a bug where we would compile `f(a, *b, c)`
as `f(a, c, *b)`.