Store line number in FuncDecl
Store the line number in FuncDecl instead of FuncIR, which allows us to
use the line number to differentiate between different functions easily
when deserializing in mypyc, when we usually only have access to a
FuncDecl.

The line number is initialized in the FuncIR constructor instead of
adding an argument to FuncDecl's constructor because there are some
places in mypyc where we create a FuncDecl before we create the
corresponding FuncIR, so we won't know the line number when the FuncDecl
is created.