Serialize line in FuncDecl
Serialize the line number in FuncDecl so that we can use it for
generating the id for this FuncDecl in get_id_from_json.

This is probably redundant as we also serialize the line number in
FuncIR, and the only time get_id_from_json is called is in module_ir.py,
where we have access to a serialized FuncIR, but storing it one more
time probably wouldn't hurt and it makes the code slightly easier to
understand.