Avoid creating OverloadedFuncDef of underscore functions
Most of mypy and mypyc assume that an OverloadedFuncDef is either an
overloaded function or a property, which becomes problematic when
non-overloads are removed in semanal or the type checker reports errors
about multiple overload definitions being required about several
potentially unrelated underscore functions. It's probably easier to just
special case underscore functions in the parser than to make all of
mypy/mypyc handle OverloadedFuncDefs of underscore functions correctly.

This also removes the check that avoids showing name redefinition errors
for underscore function OverloadedFuncDefs in semanal, as that's no
longer necessary.