Fix type signature of short_id_from_name
Change the type signature of short_id_from_name to accept None as a line
number to make it easier to pass in the _line property in FuncDecl,
which can be None if the corresponding FuncIR hasn't been created.

This also adds an assert to make sure that we don't use None as the line
number when creating the short name.