Avoid crash in register callback
Avoid trying to get the fallback type from the type variables if there
aren't 2 elements in the type variables. If there are fewer elements,
that means that we never added the fallback type to the type variables
because there was an error of some kind, in which case we should just
silently return instead of reporting more errors (especially considering
we can't do much type checking without being able to access the fallback
type).