Remove id property from FuncDef
We don't use the id property anywhere else in mypy, and adding it to
the definition of FuncDef means that we'd have to import a function from
mypyc (although we could pretty easily move that to mypy). The id might
be useful at some point for use as a unique string for identifying
functions or SymbolNodes, but for now it's easier to just determine the
id in mypyc in the one place we need it.