Check subclasses before their superclasses
When creating a dispatch function for a singledispatch function, avoid
checking any classes before we've checked all of their subclasses. That
makes sure we use the implementation with the dispatch type that appears
first in the actual argument's type's MRO, which is what the standard
library implementation of singledispatch does (it goes up the argument's
MRO, looking for classes that have registered implementations associated
with them).