Add test for singledispatch across multiple files
Add a test to make sure that we handle registering a singledispatch
function in a different module than the main function correctly.

This test uses a pattern similar to the code in
https://github.com/mypyc/mypyc/issues/802#issuecomment-883780584 to make
sure that we don't accidentally end up dynamically registering functions
in other modules and just relying on the normal singledispatch machinery
to check the dynamically registered functions.