Run register-finding pass over entire SCC
Instead of trying to look for singledispatch register implementations in
one file at a time and only using the register implementations found in
the same file as the singledispatch main function when generating code,
look for register implementations across the entire SCC once, and then
use that list when compiling every module.

That change necessitates adding an extra argument for the
PreBuildVisitor constructor so that we can remove any register calls
that we found (so we can avoid register trying to access
`__annotations__` on a builtin function) and adding an extra argument to
IRBuilder so that we can access the list of singledispatch functions
when building IR in any module.