Load functions from other modules correctly
When we're generating calls to registered implementations in other
modules, we need to load those functions by loading the module and
accessing the correct attribute, instead of trying to load that function
from the globals dict of the current module.

That change also necessitates generating imports of any modules with
registered implementations that weren't already imported so that we can
load those modules in the dispatch function.