Use native calls to registered functions
When generating calls to registered implementations in the dispatch
function for a singledispatch function, use a native call instead of a
non-native call to registered implementations to make dispatching
faster.

We still use non-native calls when the registered function has
non-register decorators, as using a native call in that case would cause
us to call the function without any modifications that the decorator
might do.