stubgen: use sys.version_info, rather than hardcoding 3.6
Not too familiar with stubgen, so maybe there's a good reason for
hardcoding, e.g. to ensure stubs capture sys.version checks in the
source? Most stubs in typeshed use sys.version checks for a) checking
Python 2, b) if they're stdlib backports. So hopefully this is a
reasonable change.

Fixes #10905