Store registry in static variable
Reusing the regular singledispatch machinery for defining registry by
applying the regular singledispatch decorator to the main function
causes every call to a singledispatch function to fail due to it not
being able to handle integers in the registry dict.

We should eventually turn that main singledispatch function into a
callable class with the necessary attributes defined, but, for now, we
store the registry in a static variable and continue to use a regular
function, which will allow us to remove the singledispatch decorator.