Preserve name of singledispatch main function
Removing the singledispatch decorator from the singledispatch main
function causes deserialization to fail because both the dispatch
function and the main function are given the same name, which the
deserialization code is not expecting.

In order to make sure that the dispatch function gets a different name
than the main function for singledispatch functions, this commit creates
a new FuncDecl that uses the mangled name for the main singledispatch
function instead of using the unmangled name of the function.