Use binary file objects with tomli
tomli recently (in 1.2.0) deprecated support for passing file objects
opened in text mode to `tomli.load`, which is causing
DeprecationWarnings in CI runs (see
https://app.travis-ci.com/github/python/mypy/jobs/529150228 for an
example).

This updates the 2 uses of tomli in mypy to open the toml file in binary
mode, and also updates the requirements for tomli to be at least 1.1.0
(because support for binary file objects was added in 1.1.0).