Compute native IDs from singledispatch_impls
Instead of assigning an ID to every registered implementation as we
generate code for each one, use builder.singledispatch_impls to
determine the IDs for all native functions. That allows us to generate
code in the dispatch function that uses the IDs of all registered native
functions, which wasn't possible with the previous implementation due to
the code for the registered implementations being generated after the
dispatch function.