Use callable class for singledispatch function
Instead of generating a regular native function for singledispatch
functions, generate a callable class. That gives us a place to put the
registry dict (instead of storing the registry dict in a static
variable) and also will allow us to support dynamically registering
functions by adding a register method to that callable class.