Add Return directly instead of using gen_return
Using gen_return in the tests that involve main singledispatch functions
that are also generators causes that gen_return call to crash, because
it tries to access the generator_class attribute.

Adding Return directly instead of going through gen_return appears to
fix the problem, probably because the dispatch function itself is never
a generator, so we don't need to support everything that gen_return
does.