Clear cache when registering new functions
Registering a new function (regardless of whether that happens in native
code or interpreted code) can cause _find_impl to return a different
result for the same input, so we should clear the cache at that point.

We might want to consider being smarter about which cache entries
actually get invalidated and which don't, but that might not be worth
the effort.