Add test
Add a test to make sure that the cache is correctly invalidated when a
function is registered dynamically from interpreted code.

This also adds a driver to the previous test that imports the
interpreted code, because previously that code wasn't getting run at
all (meaning the test was doing nothing).