[mypyc] Remove decorator_helper_name
Per mypyc/mypyc#718, renaming the original function name in the module
mapping causes the __name__ of decorated functions to have be
something like __mypyc_*_decorator_helper__. When checking if just
using the original name as the module name (and then later the function
will be overriden by the decorated value) no tests fail. And there are
indeed tests that exhibit rather complex decorator behavior so it seems
there may not actually be anything wrong with using the original name.