Now `lambda` is counted as a valid context in `handle_cannot_determine_type` (#11215)
Closes #11212

Previously `mypy` was ignoring `lambda` scope in `handle_cannot_determine_type`.
I guess it was done to fix some `lambda` related crashes.

I've also noticed one crash with this new solution in existing test. I happened because when right `or` part is unreachable, there's not `binder.frame` for it. So, I've made a temporary scope of it.