Fix cmdline package tests for namespace packages (#11323)
Prep for #9636

First test was failing because a was not a valid import (pkg.a was)

Second test  was failing because `foo.py` resolved in two ways, both
as `foo` and `foo.foo`. Disambiguating the dir from the file made
mypy path work better. The error message was very clear. Came up once
foo can be a namespace package.