Allow mypy to run self check on PyPy 3.8 beta (#11350)
A few lines change required to have mypy run on PyPy, to the extent that it 
can run self-check. Note this requires the new PyPy 3.8 beta to work.

I don't expect tests to pass because things like lxml and such are not available 
for PyPy, so I don't think we should integrate testing (yet).