Enable --fast-exit by default
This means that we kill the mypy process at exit instead of cleanly
freeing up memory. This speeds up mypy runs by ~10% in some cases
(with compiled mypy).

We've been using this internally at work for a long time without any
issues.

Previously the slowdown was visible as a small delay after the final
message from mypy was printed.

I measured the impact on a branch with some other optimizations as
well, so on master the fraction would be slightly less, as other
things would be slower.