Revert "Now `lambda` is counted as a valid context in `handle_cannot_determine_type` (#11215)" (#11694)
This reverts commit d37c2be0f7a2002a0a8c160622d937518f902cc7.

Lambdas can't be safely deferred. There are various subtle assumptions about
lambdas not being deferred in different parts of the codebase, and figuring out
how to do this safely is almost certainly not worth the trouble. This would be
very difficult to validate, since lambda deferral would probably happen very
rarely in practice.