[mypyc] Fix invalid unlikely() in certain rare branches
If we switch true/false branches, we also need to switch between
likely/unlikely.

This has an impact at least on accessing module-level final attributes
with non-constant initializers. I couldn't see a significant
difference in benchmark results on master, but this seems to help
together with some other work-in-progress improvements I have been
working on.