Run self-check on Windows as part of CI
As discussed in #11895, mypy's test suite currently does not pass a mypy self-check when run on Windows. This should hopefully be fixed by https://github.com/python/typeshed/pull/6812, but running a self-check on Windows as part of the CI tests should help avoid this issue in the future.

This new CI test should hopefully fail until #11905 is merged, confirming the diagnosis of the problem that I made in #11895. Once #11905 is merged, it should hopefully then pass, allowing this PR to be merged.