Expose modules in CheckerPluginInterface
The TypeChecker class already has the same modules attribute as
SemanticAnalyzer, but it's not exposed through the plugin interface. I
need this to call deserialize_and_fixup_type with either interface. I've
been using this with a type ignore comment in an experimental project of
mine and it seems to be working.