Add note about wrong error code in type: ignore (#12067)
If we change the error code of a message, it can result in existing
"type: ignore" comments being ignored. If this seems to be the case,
add a note that hints about changing the ignored error code.

Give a more specific message for a set of special cased error codes
where we know the original error code. In other cases give a more
vague message.