CONTRIBUTING.md: more words for first-time contributors
mypy attracts a lot of (usually student) first-time contributors who
tend to comment on issues with things like "May I work on this issue?
Can anyone provide me guidance?" and often end up not getting a
response.

The goal of this PR is to have a section we can just link to as a reply
in such situations.