Error on unused awaitable expressions
Generates an error when an expression has a type which has a
defined __await__ in its MRO but is not used. This includes all
builtin awaitable objects (e.g. futures, coroutines, and tasks)
but also anything a user might define which is also awaitable.

A hint is attached to suggest awaiting.

This can be extended in the future to other types of values that
may lead to a resource leak if needed, or even exposed as a plugin.

We test simple and complex cases (coroutines and user defined classes).
We also test that __getattr__ does not create false positives for
awaitables.

Some tests require fixes, either because they were deliberately not
awaiting an awaitable to verify some other logic in mypy, or because
reveal_type returns the object, so it was generating an error we would
rather simply silence.