stubtest: ignore more dunder positional-only errors (#12294)
This isn't actually a reversion. This logic was asymmetrical for reasons
lost to time. Although I suspect it was this way because the delta
only a few errors on typeshed.

What changed was that as a result of #12203 we actually started checking
a lot more dunder methods. Previously, we only checked dunders if either:
a) it was a special dunder, like `__init__` or `__call__`, or
b) the dunder was defined on the actual stub class itself

In particular, we started checking every dunder redefined at runtime
that the stub just inherited from object.

A possible intermediate option would be to not check positional-only
arguments in the case where a stub inherits the definition. I'll
experiment with that once typeshed CI is green.

Co-authored-by: hauntsaninja <>