Speed up caching of subtype checks
This is very performance critical. Implement a few micro-optimizations
to speed caching a bit. In particular, we use dict.get to reduce the
number of dict lookups required, and avoid tuple concatenation which
tends to be a bit slow, as it has to construct temporary objects.

It would probably be even better to avoid using tuples as keys
altogether. This could be a reasonable follow-up improvement.