[mypyc] Implement optional tracing of ops
If mypyc.irbuild.ll_builder.TRACING is set to True, print information
about ops being executed to stdout while the compiled program is run.
This can be useful for debugging and performance analysis.

Only those ops generated in the main irbuild pass which have a line
number will be included in the trace. This seems sufficient to make
this feature useful, but it may make sense to include additional ops
in the future.

I've already used this to successfully fix a tricky code generation
bug.

Example trace from a simple program (it prints `hello` and appends
to a list):
```
[trace] t:9::Call(bar)
[trace] t:6:bar:CallC(CPyObject_GetAttr)
[trace] t:6:bar:CallC(_PyObject_Vectorcall)
hello
[trace] t:7:bar:Call(foo)
[trace] t:2:foo:CallC(PyList_New)
[trace] t:3:foo:CallC(PyList_Append)
[trace] t:3:foo:ComparisonOp
```