make_simplified_union: add caching and reduce allocations
make_simplified_union is used in a lot of places and therefore
accounts for a significant share to typechecking time. Based
on sample metrics gathered from a large real-world codebase
we can see that:
 1. the majority of inputs are already as simple as they're
    going to get, which means we can avoid allocation extra
    lists and return the input unchanged
 2. most of the cost of `make_simplified_union` comes from
    `is_proper_subtype`
 3. `is_proper_subtype` has some caching going on under the hood
    but it only applies to `Instance`, and cache hit rate is low
    in this particular case because, as per 1) above, items are
    in fact rarely subtypes of each other

To address 1, refactor `make_simplified_union` with an optimistic
fast path that avoid unnecessary allocations.

To address 2 & 3, introduce a cache to record the result of union
simplification.

These changes are observed to yield significant improvements in
a real-world codebase: a roughly 10-20% overall speedup, with
make_simplified_union/is_proper_subtype no longer showing up as
hotspots in the py-spy profile.

For #12526