Automatically detect decorator argument default.
Instead of relying only on hardcoded default value, first try to find
out the actuall default value of a bool argument to a the
class/attrib-maker decorators, and only if the value could not be
extracted from the code, fall back to the default value defined in mypy.

This enables the correct handling of decorators which wrap attrs
decorators but set different defaults, for example to kw_only.

This commit also adds to test cases. testAttrsWrappedDecorators tests
the existing support for wrapping attrs decorators, via a "fake plugin"
and testAttrsWrappedDecoratorsWithDifferentDefaults tests the new
support for decorators that have different default values than the attrs
decorators.