Pass on unexpanded type aliases
As recommended in the docstring for get_proper_type, when you call
another function after expanding a type alias, pass the original
unexpanded type alias on instead of the expanded version.

This tries to fix a few places where we don't follow that in the error
formatting code so that unexpanded aliases make it to the error
formatting code.