Expand aliases if necessary in unfollowed import errors
If the underlying type that an alias refers to has part of it turned to
Any due to an unfollowed import, expand the alias before printing it in
the error to make it obvious where the Any actually shows up.

We avoid this in the case of the Any due to unfollowed import showing up
in the generic arguments, because then the error printing code will
clearly print the argument as Any.