Put filelock around PEP 561 tests (fixes #12615) (#12857)
Fixes #12615.

As discussed in the linked issue, this PR would put a filesystem-based lock around the pip 
install steps of the suite of PEP561 testcases, to address race conditions.

It introduces a test-dependency on tox-dev/py-filelock. I used 3.0.0 as a lower bound, simply 
because that's what the tox version mentioned in tox.ini specifies as its lower bound. 
However from the release history it seems that Python 3.6 support was dropped in 3.4.2, 
hence the second line in the requirements.

I ended up just adding the location of the lock file to .gitignore -- I guess it would be 
possible to try and remove it after the suite has run, e.g. with atexit.register, but I didn't 
want to make this complicated.