[mypyc] Support native int argument defaults
Since native ints can't have a dedicated reserved value to mark a
missing argument, use extra bitmap arguments that are used to indicate
whether default values should be used for particular arguments. The
bitmap arguments are implicitly added at the end of a signature. They
are not visible to user code and are just an implementation detail.

This is analogous to how we track definedness of native int attributes
using bitmaps.

If a function accepts no arguments with a native int type and a default
value, the bitmap arguments won't be generated.

This doesn't handle inheritance properly. I'll fix inheritance in a
follow-up PR.

Work on mypyc/mypyc#837.