[mypyc] Native int primitives (#12973)
Add various C primitives that will be used to support native ints.

The primitives aren't used for anything yet. I'll prepare follow-up PRs
that use the primitives and include tests. I'm splitting these into a
separate PR to make this easier to review. All of these are tested in
my local branch, at least to a basic level.

Most of these are fairly straightforward, but we need to jump through
some hoops to make the semantics of // and % operators compatible
with Python semantics when using negative operands.

Work on mypyc/mypyc#837.