[mypyc] Add LoadAddress op for PySet_Type & PyFrozenSet_Type
This also fixes https://github.com/mypyc/mypyc/issues/917

RE above, the root issue is that mypyc didn't know builtins.set was a
built-in name, so it guessed it comes from the module globals. This
didn't blow up anything up somehow... until the dataclasses
commit[^1] which made the `__annotations__` logic for dataclasses try to
better preserve the type annotations (previously they would be erased
to builtins.type). This new logic would use `load_type` to load
`builtins.set` (so it can be put in `__annotations__`) which went poorly
as only types registered with `load_address_op` are considered
built-ins.

[^1]: https://github.com/python/mypy/commit/1bcfc041bb767ee93e90676b0a61f3e40267e858