stubtest: don't error for a missing submodule if the module name is private
If a module name is private, the contents of that module are likely considered implementation details by the package author. Sometimes we want to include these implementation details in a stub, and sometimes we don't, but in my opinion stubtest probably shouldn't error for these submodules being missing from the stub. This is in keeping with stubtest's approach for class and module attributes, where it generally doesn't emit an error if the attribute name is private.